added balance-timeline component
